requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
Babel==2.12.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==2.4.0
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
daphne==2.5.0
dash==1.20.0
dash-bootstrap-components==0.13.1
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
dataclasses-json==0.5.9
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.19
django-admin-soft-dashboard==1.0.12
django-bootstrap-v5==1.0.11
django-bootstrap4==1.1.1
django-bootstrap5==23.1
django-neomodel==0.0.7
django-plotly-dash==1.7.2
docutils==0.18.1
dominate==2.8.0
dpd-components==0.1.0
dpd-static-support==0.0.5
fastjsonschema==2.17.1
Flask==2.1.3
Flask-Compress==1.13
fonttools==4.40.0
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
h5py==3.9.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.7.0
importlib-resources==6.0.1
incremental==22.10.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-client==8.3.0
jupyter-core==5.3.1
jupyterlab-pygments==0.2.2
keyring==24.2.0
kiwisolver==1.4.4
langchain==0.0.222
langchainplus-sdk==0.0.20
lasso-python==2.0.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
martor==1.6.7
matplotlib==3.7.2
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nbsphinx==0.9.2
neo4j==5.10.0
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.0.1
networkx==3.1
numexpr==2.8.4
numpy
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.10
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
rdflib==6.3.2
recommonmark==0.7.1
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==12.6.0
rpds-py==0.8.8
scikit-learn==1.3.0
scipy
SecretStorage==3.3.3
service-identity==23.1.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.4.1
sphinx==6.2.1
sphinx-mdinclude==0.5.3
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.17
sqlparse==0.4.4
streamlit==1.24.0
streamlit-chat==0.1.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.5.0
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
